+++ lib/core/router/route_paths.dart
@@ -1,4 +1,5 @@
 class RoutePaths {
+  static const String exerciseAdd = 'exercise/add';
   static const String dashboard = 'dashboard';
   static const String home = '/';
   static const String onboardingWelcome = 'onboarding/welcome';