+++ lib/core/router/router.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:power_progress/features/exercise/presentation/pages/exercise/exercise_form_page.dart';
 
 import '../../features/exercise/presentation/pages/dashboard/dashboard_page.dart';
 import '../../features/exercise/presentation/pages/onboarding/onboarding_exercise_page.dart';
@@ -18,6 +19,14 @@ class Router {
           builder: (_) => DashboardPage(),
           settings: const RouteSettings(
             name: RoutePaths.dashboard,
+          ),
+        );
+
+      case RoutePaths.exerciseAdd:
+        return MaterialPageRoute(
+          builder: (_) => ExerciseFormPage(),
+          settings: const RouteSettings(
+            name: RoutePaths.exerciseAdd,
           ),
         );