+++ lib/features/exercise/data/datasources/i_exercise_datasource.dart
@@ -1,0 +1,6 @@
+import '../models/exercise_model.dart';
+
+abstract class IExerciseDatasource {
+  Future<ExerciseModel> add(ExerciseModel exercise);
+  Future<List<ExerciseModel>> get();
+}