+++ lib/features/exercise/domain/repositories/i_exercise_repository.dart
@@ -1,0 +1,6 @@
+import '../entities/exercise.dart';
+
+abstract class IExerciseRepository {
+  Future<Exercise> add(Exercise exercise);
+  Future<List<Exercise>> get();
+}