+++ lib/features/exercise/presentation/bloc/exercise_event.dart
@@ -2,4 +2,18 @@ part of 'exercise_bloc.dart';
 
 abstract class ExerciseEvent extends Equatable {
   const ExerciseEvent();
+
+  @override
+  List<Object> get props => [];
 }
+
+class ExerciseAddEvent extends ExerciseEvent {
+  final Exercise exercise;
+
+  const ExerciseAddEvent({@required this.exercise});
+
+  @override
+  List<Object> get props => [exercise];
+}
+
+class ExerciseGetEvent extends ExerciseEvent {}